Ignore request for mutexes
Mutex support is missing in this implementation, but this prevents pk11-kit from working.
pk11-kit is commonly used to set up pkcs11 in Debian.

For now we can ignore the mutex requests, as we do not anticipate issues with concurrency between
pkcs11 modules, and the SGX implementation is meant to be thread safe.